partition、assigment非多对多的细致权限

+++ partition/views/domain.py
@@ -1,4 +1,4 @@
-from op_keystone.exceptions import CustomException
+from op_keystone.exceptions import CustomException, PermissionDenied
 from op_keystone.base_view import BaseView
 from utils import tools
 from utils.dao import DAO
@@ -13,11 +13,17 @@ class DomainsView(BaseView):
 
     def get(self, request):
         try:
-            # 获取 uuid 参数，若有则为单个对象获取
+            # 非云管理员直接返回当前 domain 信息
+            if not request.cloud_admin:
+                obj = self.domain_model.get_obj(uuid=request.user.domain)
+                return self.standard_response(obj.serialize())
+
+            # 提取 uuid 参数
             uuid_opts = ['uuid']
             request_params = self.get_params_dict(request, nullable=True)
             uuid_opts_dict = self.extract_opts(request_params, uuid_opts, necessary=False)
 
+            # 若存在 uuid 参数则返回获取的单个对象
             if uuid_opts_dict:
                 obj = self.domain_model.get_obj(**uuid_opts_dict)
                 return self.standard_response(obj.serialize())
@@ -26,10 +32,11 @@ class DomainsView(BaseView):
             page_opts = ['page', 'pagesize']
             page_opts_dict = self.extract_opts(request_params, page_opts, necessary=False)
 
-            # 当前页数据列获取
+            # 当前页数据获取
             total_list = self.domain_model.get_dict_list()
             page_list = tools.paging_list(total_list, **page_opts_dict)
 
+            # 返回数据
             return self.standard_response(page_list)
 
         except CustomException as e:
@@ -37,32 +44,50 @@ class DomainsView(BaseView):
 
     def post(self, request):
         try:
-            # 参数提取
+            # 非云管理员直接进行权限拒绝
+            if not request.cloud_admin:
+                raise PermissionDenied()
+
+            # 定义参数提取列表
             necessary_opts = ['name', 'company']
-            extra_opts = ['enable', 'comment']
+            extra_opts = ['is_main', 'enable', 'comment']
+
+            # 参数提取
             request_params = self.get_params_dict(request)
             necessary_opts_dict = self.extract_opts(request_params, necessary_opts)
             extra_opts_dict = self.extract_opts(request_params, extra_opts, necessary=False)
 
-            # 参数合成
-            obj_field = {}
+            # 参数合成，预设 create_by 的值
+            obj_field = {
+                'created_by': request.user.uuid
+            }
             obj_field.update(necessary_opts_dict)
             obj_field.update(extra_opts_dict)
-            obj_field['created_by'] = request.user.uuid
 
-            # 创建对象
-            ins = self.domain_model.create_obj(**obj_field)
+            # 对象创建
+            check_methods = ('pre_save',)
+            obj = self.domain_model.create_obj(check_methods=check_methods, **obj_field)
 
-            return self.standard_response(ins.serialize())
+            # 返回创建的对象
+            return self.standard_response(obj.serialize())
 
         except CustomException as e:
             return self.exception_to_response(e)
 
     def put(self, request):
         try:
-            # 参数提取
+            # 非云管理员直接进行权限拒绝
+            if not request.cloud_admin:
+                raise PermissionDenied()
+
+            # 定义参数提取列表
             necessary_opts = ['uuid']
-            extra_opts = ['name', 'purpose', 'enable', 'comment']
+            extra_opts = [
+                'name', 'company', 'enable',
+                'is_main', 'comment'
+            ]
+
+            # 参数提取
             request_params = self.get_params_dict(request)
             necessary_opts_dict = self.extract_opts(request_params, necessary_opts)
             extra_opts_dict = self.extract_opts(request_params, extra_opts, necessary=False)
@@ -71,9 +96,11 @@ class DomainsView(BaseView):
             obj = self.domain_model.get_obj(**necessary_opts_dict)
 
             # 对象更新
+            check_methods = ('pre_save',)
             extra_opts_dict['updated_by'] = request.user.uuid
-            updated_obj = self.domain_model.update_obj(obj, **extra_opts_dict)
+            updated_obj = self.domain_model.update_obj(obj, check_methods=check_methods, **extra_opts_dict)
 
+            # 返回更新的对象
             return self.standard_response(updated_obj.serialize())
 
         except CustomException as e:
@@ -81,14 +108,21 @@ class DomainsView(BaseView):
 
     def delete(self, request):
         try:
-            # 参数获取
+            # 非云管理员直接进行权限拒绝
+            if not request.cloud_admin:
+                raise PermissionDenied()
+
+            # 定义参数提取列表
             necessary_opts = ['uuid']
+
+            # 参数获取
             request_params = self.get_params_dict(request)
             necessary_opts_dict = self.extract_opts(request_params, necessary_opts)
 
             # 对象获取
             deleted_obj = self.domain_model.delete_obj(**necessary_opts_dict)
 
+            # 返回成功删除
             return self.standard_response('succeed to delete %s' % deleted_obj.name)
 
         except CustomException as e:
